remove rc-slider hook
